+++ documentation/Changelog.md
@@ -1,4 +1,5 @@
 # Changelog
+* [1.4.6](#1.4.6) Released on 13 june 2016 
 * [1.4.5](#1.4.5) Released on 18 mai 2016 
 * [1.4.4](#1.4.4) Released on 04 mai 2016 
 * [1.4.3](#1.4.3) Released on 22 april 2016 
@@ -7,6 +8,13 @@
 * [1.4.0](#1.4.0) Released on 31 march 2016
 
 ---
+
+## 1.4.6
+* [CONJ-293] Permit named pipe connection without host
+* [CONJ-309] Possible NPE on aurora when failover occur during connection initialisation
+* [CONJ-312] NPE while loading a null from TIMESTAMP field using binary protocol
+* [misc] batch with one parameter correction (using rewriteBatchedStatements option)
+
 ## 1.4.5
 * [CONJ-297] Useless memory consumption when using Statement.setQueryTimeout
 * [CONJ-294] PrepareStatement on master reconnection after a failover